+++ packages/apps/dev-wallet/src/App/app.tsx
@@ -3,12 +3,13 @@ import { LayoutProvider } from '@/modules/layout/layout.provider.tsx';
 import { WalletProvider } from '@/modules/wallet/wallet.provider';
 import { useTheme } from '@kadena/kode-ui';
 import { PromptProvider } from '../Components/PromptProvider/Prompt';
+import { BetaHeader } from './BetaHeader';
 import { Routes } from './routes';
 import { SessionProvider } from './session';
 
 function Providers({ children }: { children: React.ReactNode }) {
   // initialize the theme
-  useTheme();
+  useTheme({ overwriteTheme: 'dark' });
   return (
     <SessionProvider>
       <DatabaseProvider>
@@ -24,6 +25,7 @@ function Providers({ children }: { children: React.ReactNode }) {
 
 export const App = () => (
   <Providers>
+    <BetaHeader />
     <Routes />
   </Providers>
 );