+++ packages/apps/tools/src/pages/modules/explorer/index.tsx
@@ -87,6 +87,7 @@ const ModuleExplorerPage = (
 
   const mainnetModulesQuery = useModulesQuery('mainnet01');
   const testnetModulesQuery = useModulesQuery('testnet04');
+  const testnet05ModulesQuery = useModulesQuery('testnet05');
 
   const setDeepLink = useCallback(
     (module: IncompleteModuleModel) => {
@@ -137,6 +138,13 @@ const ModuleExplorerPage = (
             key: 'testnet04',
             data: testnetModulesQuery.data!,
             isLoading: testnetModulesQuery.isFetching,
+            supportsReload: true,
+          },
+          {
+            title: 'Testnet05',
+            key: 'testnet05',
+            data: testnet05ModulesQuery.data!,
+            isLoading: testnet05ModulesQuery.isFetching,
             supportsReload: true,
           },
         ]}