+++ packages/libs/snap/src/services/getActiveNetwork.ts
@@ -1,0 +1,5 @@
+import { ApiParams } from '../types';
+
+export const getActiveNetwork = (snapApi: ApiParams): string => {
+  return snapApi.state.activeNetwork;
+};