+++ packages/libs/snap/src/utils/constants.ts
@@ -1,0 +1,2 @@
+export const BASE_ACCOUNT_NAME = 'Kadena Account';
+export const DEFAULT_CHAIN_PREFERENCES = { 0: 0, 1: 1 };